backend/requirements.txt
# AtlasField Backend

fastapi==0.128.0
uvicorn[standard]==0.40.0
pydantic==2.12.5
pydantic-settings==2.12.0
sqlalchemy[asyncio]==2.0.46
asyncpg==0.31.0
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.22

# Sentinel Hub API
sentinelhub==3.11.3
numpy==1.26.4

# Gemini API
google-generativeai==0.8.6

# Utilities
httpx==0.28.1
python-dotenv==1.2.1
